- Added the dbal layer to the console commands so that they'll automatically use an aurora database if the app is configured to do so - Updated the sls_app.yaml file to only have the single symfony console command as a lambda function, and different arguments are passed to it like a normal symfony console app - Added a composer script to initialize the database schema for aurora - Updated the deployment.md to be slightly more correct and make a note of the serverless nature of the database

+++ api/src/App/Commands/InitDB.php
@@ -5,13 +5,17 @@ namespace TOE\App\Commands;
 
 
 use Aws\RDSDataService\Exception\RDSDataServiceException;
+use Aws\RDSDataService\RDSDataServiceClient;
 use Aws\S3\Exception\S3Exception;
 use Silly\Input\InputArgument;
 use Silly\Input\InputOption;
 use Symfony\Component\Console\Helper\FormatterHelper;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use TOE\App\Service\AWS\AuroraDataAPIWrapper;
+use TOE\App\Service\AWSConfig;
 use TOE\GlobalCode\Constants;
+use TOE\GlobalCode\Env;
 
 class InitDB extends aCmd
 {
@@ -21,32 +25,49 @@ class InitDB extends aCmd
 	{
 		$this->setDescription('Initializes the mysql database to be ready for use for trick-or-eat');
 		$this->addArgument('querydir', InputArgument::OPTIONAL, 'A path to a directory that holds .sql files used to initialize the database. __DIR__ will be replaced by ' . __DIR__, __DIR__ . '/../../../../.docker/mysql/data');
-		$this->addOption('aurora', 'aur', InputOption::VALUE_NONE, 'Use this when populating an aurora serverless database');
 		$this->addOption('wipe', null, InputOption::VALUE_NONE, 'Use this to wipe the contents of the existing database if there is one');
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output)
 	{
 		$wipe = $input->getOption('wipe');
-		if($input->getOption('aurora'))
+		try
 		{
-			return $this->initAurora($input, $output, $wipe);
+			return $this->init($input, $output, $wipe);
+		}
+		catch(\Exception $ex)
+		{
+			if(stripos($ex->getMessage(), "The last packet sent successfully to the server was 0 milliseconds ago") !== false)
+			{
+				$output->writeln("Aurora Serverless might be warming up. Check the current capacity size of the cluster");
+				return 1;
+			}
+			else
+			{
+				throw $ex;
+			}
+		}
+
+	}
+
+	protected function init(InputInterface $input, OutputInterface $output, bool $wipe)
+	{
+		//check to see if the TOE database exists, and if it does, print and exit
+		$this->notifyUserOfDBType($output);
+		if (!$this->isAurora())
+		{
+			$this->resetDbConn();
 		}
 		else
 		{
-			return $this->initMysql($input, $output, $wipe);
+			$this->createAuroraSchema(Constants::DATABASE_NAME);
 		}
-	}
-
-	protected function initMysql(InputInterface $input, OutputInterface $output, bool $wipe)
-	{
-		//check to see if the TOE database exists, and if it does, print and exit
-		$schemas = $this->container->db->query("SHOW DATABASES");
+		$schemas = $this->container->dbConn->query("SHOW DATABASES")->fetchAll();
 		if ($wipe)
 		{
 			try
 			{
-				$this->container->db->query("DROP DATABASE " . Constants::DATABASE_NAME . ";");
+				$this->container->dbConn->query("DROP DATABASE " . Constants::DATABASE_NAME . ";");
 			}
 			catch(\Exception $ex)
 			{
@@ -59,7 +80,13 @@ class InitDB extends aCmd
 					throw $ex;
 				}
 			}
-
+		}
+		else
+		{
+			if ($this->checkAuroraDBAlreadyInit($output) !== false)
+			{
+				return 1;
+			}
 		}
 
 		if(!empty($schemas))
@@ -74,8 +101,8 @@ class InitDB extends aCmd
 			}
 			if($hasTOE)
 			{
-				$tables = $this->container->db->query("SELECT DISTINCT table_name FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = '" . Constants::DATABASE_NAME . "'");
-				if (!empty($tables))
+				$tables = $this->container->dbConn->query("SELECT DISTINCT table_name FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = '" . Constants::DATABASE_NAME . "'")->fetchAll();
+				if (!empty($tables) && !$wipe)
 				{
 					$output->writeln("'toe' database already exists!");
 					$output->writeln("Ensure the data is backed up and delete the schema with 'DROP DATABASE " . Constants::DATABASE_NAME . "'.");
@@ -86,60 +113,15 @@ class InitDB extends aCmd
 			}
 		}
 
-		$dir = $input->getArgument('querydir');
-		$dir = str_replace('__DIR__', __DIR__, $dir);
-		$dir = rtrim($dir, "/");
-		//run the mysql dump commands
-		$files = array_diff(scandir($dir), ['.', '..']);
-		foreach($files as $file)
-		{
-			$query = file_get_contents($dir . "/" . $file);
-			$affected = $this->container->db->rawExecuteNonQuery($query);
-			$output->writeln("Finished running $file, $affected rows affected");
-		}
-
-		$res = $this->container->db->query("SELECT COUNT(DISTINCT table_name) as cnt FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = '" . Constants::DATABASE_NAME . "'");
-		$tableCount = $res[0]['cnt'];
-		$output->writeln("Database " . Constants::DATABASE_NAME . " now has $tableCount tables");
-
-
-		return 0;
-	}
-
-	protected function initAurora(InputInterface $input, OutputInterface $output, bool $wipe)
-	{
-		//TODO: implement the wipe option to reset the database for tests
-		if ($wipe)
-		{
-			try
-			{
-				$this->container->aurora->queryDB("DROP DATABASE " . Constants::DATABASE_NAME . ";");
-			}
-			catch(\Exception $ex)
-			{
-				$output->writeln("Error while trying to wipe database: " . $ex->getMessage());
-				$output->writeln("Exiting...");
-				return 1;
-			}
-		}
-		if ($this->checkAuroraDBAlreadyInit($output) !== false)
-		{
-			return 1;
-		}
-
-		$dir = $input->getArgument('querydir');
-		$dir = str_replace('__DIR__', __DIR__, $dir);
-		$dir = rtrim($dir, "/");
-		//run the mysql dump commands
-		$files = array_diff(scandir($dir), ['.', '..']);
-		if (empty($files))
-		{
-			$output->writeln("Unable to find any files to initialize the database with");
-			return 1;
-		}
 		/** @var FormatterHelper $formatter */
 		$formatter = $this->getHelper('formatter');
-		$this->container->aurora->executeStatement("SET FOREIGN_KEY_CHECKS=0");
+
+		$this->container->dbConn->exec("SET FOREIGN_KEY_CHECKS=0");
+		$dir = $input->getArgument('querydir');
+		$dir = str_replace('__DIR__', __DIR__, $dir);
+		$dir = rtrim($dir, "/");
+		//run the mysql dump commands
+		$files = array_diff(scandir($dir), ['.', '..']);
 		foreach($files as $file)
 		{
 			$query = file_get_contents($dir . "/" . $file);
@@ -157,13 +139,17 @@ class InitDB extends aCmd
 				}
 				$output->writeln(sprintf("Executing statement %d", $idx + 1));
 				$output->writeln($singleQuery);
-				$res = $this->container->aurora->executeStatement($singleQuery);
-				$affected += $res->get('numberOfRecordsUpdated');
+				$affected = $this->container->dbConn->exec($singleQuery);
 			}
 
 			$output->writeln("Finished running $file, $affected rows affected");
 		}
-		$this->container->aurora->executeStatement("SET FOREIGN_KEY_CHECKS=1");
+
+		$res = $this->container->dbConn->query("SELECT COUNT(DISTINCT table_name) as cnt FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = '" . Constants::DATABASE_NAME . "'")->fetch();
+		$this->container->dbConn->exec("SET FOREIGN_KEY_CHECKS=1");
+		$tableCount = $res['cnt'];
+		$output->writeln("Database " . Constants::DATABASE_NAME . " now has $tableCount tables");
+
 
 		return 0;
 	}
@@ -171,29 +157,12 @@ class InitDB extends aCmd
 	protected function checkAuroraDBAlreadyInit(OutputInterface $output)
 	{
 		//check to see if the TOE database exists, and if it does and has tables, print and exit
-		try
-		{
-			$res = $this->container->aurora->executeStatement("SHOW DATABASES");
-		}
-		catch(RDSDataServiceException $ex)
-		{
-			if(stripos($ex->getMessage(), "The last packet sent successfully to the server was 0 milliseconds ago. The driver has not recieved any packets from the server") !== false)
-			{
-				$output->writeln("Aurora Serverless might be warming up. Check the current capacity size of the cluster");
 
-				return true;
-			}
-			else
-			{
-				throw $ex;
-			}
-		}
-		$schemas = $res->get('records');
-
+		$res = $this->container->dbConn->query("SHOW DATABASES");
 		$hasTOE = false;
-		foreach($schemas as $schemaRow)
+		foreach($res as $database)
 		{
-			if($schemaRow[0]['stringValue'] === Constants::DATABASE_NAME)
+			if($database['Database'] === Constants::DATABASE_NAME)
 			{
 				$hasTOE = true;
 			}
@@ -201,9 +170,8 @@ class InitDB extends aCmd
 		if($hasTOE)
 		{
 			//look for the tables
-			$res = $this->container->aurora->executeStatement("SELECT DISTINCT table_name FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = '" . Constants::DATABASE_NAME . "'");
-			$tables = $res->get('records');
-			if(!empty($tables))
+			$res = $this->container->dbConn->query("SELECT DISTINCT table_name FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = '" . Constants::DATABASE_NAME . "'");
+			if(!empty($res))
 			{
 				$output->writeln("'toe' database already has tables!");
 				$output->writeln("Ensure the data is backed up and delete the schema with 'DROP DATABASE " . Constants::DATABASE_NAME . "'.");
@@ -214,5 +182,26 @@ class InitDB extends aCmd
 		}
 
 		return false;
+	}
+
+	protected function resetDbConn()
+	{
+		$configs = $this->container->getDbConfigs();
+		//for when connecting to regular mysql and you've wiped out the toe schema...
+		if (isset($configs['dbname']))
+		{
+			unset($configs['dbname']);
+		}
+		$this->container->setNewConnection($configs);
+	}
+
+	protected function createAuroraSchema(string $schemaName)
+	{
+		$awsConfigs = AWSConfig::getStandardConfig();
+		$awsConfigs['version'] = '2018-08-01';
+		$aurora = (new AuroraDataAPIWrapper(new RDSDataServiceClient($awsConfigs)))
+			->setDbArn(Env::get(Env::TOE_DB_ARN))
+			->setSecretArn(Env::get(Env::TOE_DB_SECRET_ARN));
+		$aurora->queryDB("CREATE SCHEMA toe;");
 	}
 }
